[FLINK-30288] Use PredicateVisitor to convert predicate for orc
